加入道路宽度补线，出环变得稳定了很多

+++ SMARTCAR/project/code/image.h
@@ -50,9 +50,9 @@ extern uint8 binary_image[MT9V03X_H][MT9V03X_W];
 
 extern uint8 image_threshold;
 
-extern uint8 left_line[MT9V03X_H];
+extern int16 left_line[MT9V03X_H];
 
-extern uint8 right_line[MT9V03X_H];
+extern int16 right_line[MT9V03X_H];
 
 extern float line_err;
